Add more refactor task to TODO. Moved everything UI related to ui.go
